@@ -1,4 +1,4 @@
-class CreateAppointments < ActiveRecord::Migration
+class CreateAppointments < ActiveRecord::Migration[4.2]
   def change
     create_table :appointments do |t|
       t.datetime :scheduled_at
